// source/client/routes/static/v1/routes/static/opportunities/methods/body-of-request-to-post/BodyOfRequestToPostOfMethodsOfOpportunitiesOfStaticOfRoutesOfV1OfStaticOfRoutesOfClient.ts
export type BodyOfRequestToPostOfMethodsOfOpportunitiesOfStaticOfRoutesOfV1OfStaticOfRoutesOfClient =
	{
		readonly file: null | Uint8Array<ArrayBuffer>;
		readonly fileName: null | string;
		readonly fileType: `Image` | `Pdf`;
		readonly metadata: null | {readonly [key: string]: null | string};
		readonly type: `Cmrs` | `Fines` | `Insurances` | `Invoices` | `Receipts`;
	};